split some more
