+++ tests/playwright/sanity/modules/nested-tabs/nested-tabs-5.test.ts-snapshots/tabs-direction-top-icon-position-right-align-left-linux.jpeg
